Move ELA.Utility dependency to ELA/utils
