+++ config.go
@@ -11,7 +11,7 @@ import (
 	"github.com/elastos/Elastos.ELA.SideChain.ID/params"
 
 	"github.com/elastos/Elastos.ELA/common"
-	"github.com/elastos/Elastos.ELA.Utility/elalog"
+	"github.com/elastos/Elastos.ELA/utils/elalog"
 )
 
 const (
